+++ backend/src/main/java/com/youruber/youruber/respositories/PassageiroRepository.java
@@ -1,11 +1,15 @@
 package com.youruber.youruber.respositories;
+
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.youruber.youruber.entities.Passageiro;
 
 @Repository
-public interface PassageiroRepository extends JpaRepository<Passageiro, Long> {
+public interface PassageiroRepository extends JpaRepository<Passageiro, Integer> {
+
+	
 
 	
 }
